requirements.txt
yt-dlp>=2023.12.30
duckduckgo-search>=4.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
ffmpeg-python>=0.2.0
httpx>=0.25.0
